pyproject.toml
[tool.poetry]
name = "appchat"
version = "0.1.0"
description = ""
authors = ["Helder Ferreira Fernandes <geral.helder@outlook.pt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
discord-py = "^2.3.0"
python-dotenv = "^1.0.0"
openai = "^0.27.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
